refactor tests into a separate crate
