+++ Cargo.toml
@@ -62,6 +62,7 @@ members = [
   "token/transfer-hook/interface",
   "token/confidential-transfer/proof-extraction",
   "token/confidential-transfer/proof-generation",
+  "token/confidential-transfer/proof-tests",
   "token/client",
   "utils/cgen",
   "utils/test-client",
