+++ token/confidential-transfer/proof-extraction/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "spl-token-confidential-transfer-proof-extraction"
 version = "0.1.0"
-description = "Solana Program Library Confidential Transfer Proofs"
+description = "Solana Program Library Confidential Transfer Proof Extraction"
 authors = ["Solana Labs Maintainers <maintainers@solanalabs.com>"]
 repository = "https://github.com/solana-labs/solana-program-library"
 license = "Apache-2.0"
@@ -10,9 +10,6 @@ edition = "2021"
 [dependencies]
 solana-zk-sdk = "2.0.0"
 thiserror = "1.0.61"
-
-[dev-dependencies]
-spl-token-confidential-transfer-proof-generation = { version = "0.1.0", path = "../proof-generation" }
 
 [lib]
 crate-type = ["cdylib", "lib"]
