+++ token/confidential-transfer/proof-generation/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "spl-token-confidential-transfer-proof-generation"
 version = "0.1.0"
-description = "Solana Program Library Confidential Transfer Proofs"
+description = "Solana Program Library Confidential Transfer Proof Generation"
 authors = ["Solana Labs Maintainers <maintainers@solanalabs.com>"]
 repository = "https://github.com/solana-labs/solana-program-library"
 license = "Apache-2.0"
