+++ token/confidential-transfer/proof-tests/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "spl-token-confidential-transfer-proof-test"
+version = "0.0.1"
+description = "Solana Program Library Confidential Transfer Proof Test"
+authors = ["Solana Labs Maintainers <maintainers@solanalabs.com>"]
+repository = "https://github.com/solana-labs/solana-program-library"
+license = "Apache-2.0"
+edition = "2021"
+
+[dev-dependencies]
+curve25519-dalek = "3.2.1"
+solana-zk-sdk = "2.0.0"
+thiserror = "1.0.61"
+spl-token-confidential-transfer-proof-extraction = { version = "0.1.0", path = "../proof-extraction" }
+spl-token-confidential-transfer-proof-generation = { version = "0.1.0", path = "../proof-generation" }
